# ONGConnect/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ONGConnect — Match Voluntários e ONGs",
        "description": "API REST para conexão entre ONGs do Alto Vale do Itajaí e voluntários da comunidade. Projeto de Extensão — Unidavi 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/match/sugestoes": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Sugere demandas abertas rankeadas por score de match para o voluntário autenticado",
                "operationId": "b344e7d575544b4e63af121e0054165c",
                "parameters": [
                    {
                        "name": "raio_km",
                        "in": "query",
                        "description": "Raio máximo de busca em km (padrão: 50)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 50
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de sugestões ordenadas por score",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "demanda": {
                                                        "type": "object"
                                                    },
                                                    "score": {
                                                        "properties": {
                                                            "total": {
                                                                "type": "number",
                                                                "example": 0.78
                                                            },
                                                            "categoria": {
                                                                "type": "number",
                                                                "example": 0.8
                                                            },
                                                            "proximidade": {
                                                                "type": "number",
                                                                "example": 0.75
                                                            },
                                                            "distancia_km": {
                                                                "type": "number",
                                                                "example": 12.5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Apenas voluntários recebem sugestões"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/match/score": {
            "get": {
                "tags": [
                    "Match"
                ],
                "summary": "Retorna o score de match entre o voluntário autenticado e uma demanda específica. ONGs veem os scores de todos os seus inscritos.",
                "operationId": "cd4cdeac31a21327ae5dc0e756437a41",
                "parameters": [
                    {
                        "name": "demanda_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Score detalhado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Demanda não encontrada"
                    },
                    "422": {
                        "description": "Parâmetro demanda_id ausente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Autenticação via token Bearer (Laravel Sanctum)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}